# docs/swagger.yaml
definitions:
  request.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 1
        type: string
    required:
    - email
    - password
    type: object
  request.RegisterRequest:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
    required:
    - confirm_password
    - email
    - password
    type: object
  response.GetUserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
    type: object
  response.LoginResponse:
    properties:
      token:
        type: string
    type: object
info:
  contact: {}
paths:
  /db-dashboards/api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: login user via JWT
      parameters:
      - description: login info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LoginResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login user
      tags:
      - Auth
  /db-dashboards/api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: to register new user
      parameters:
      - description: registration info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RegisterRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GetUserResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register new user
      tags:
      - Auth
swagger: "2.0"